apps/asmqttviewer/requirements.txt
ansicolors
paho-mqtt